Add non-interactive -u|--upgrade option
When creating new docker images based on the worker image, it is useful
to be able to ensure the latest version of prepare_meta is installed.
Unfortunately, the current version prompts the user if an update is
found. This makes it hard to automatically build derived images.

To avoid this, a new option is added: -u|--upgrade

When specified, the script simply ensures that the latest version
is downloaded and then exits.

All other options, with the exception of -h|--help, are ignored.

Thus a docker image can contain the statements

  RUN prepare_meta -u && prepare_meta -f some_new_flavor ...

allowing a flavor, not present in script contained in the official
container, but known to exist in a newer version, to be safely
specified.

Change-Id: I973c7dcf6310c8189ebd90f7a4683e3751160b64
Signed-off-by: Keith Derrick <keith.derrick@lge.com>